requirements.txt
asgiref==3.5.0
Django==4.0.4
djangorestframework==3.13.1
mypy==0.942
mypy-extensions==0.4.3
Pillow==9.1.0
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0